requirements.txt
Django>=4.2.0
channels>=4.0.0
channels-redis>=4.1.0
google-generativeai>=0.3.0
google-adk>=0.1.0
PyJWT>=2.8.0
python-multipart>=0.0.6
daphne>=4.0.0
redis>=5.0.0
pydub>=0.25.1

